Don't hardcode /usr/lib/perl5 in debian/libwin-hivex-perl.install (Closes: #752015)
Starting with version 5.20.0, the Debian perl package is changing
the "vendorarch" library paths (currently /usr/lib/perl5) to include
the multiarch triplet and the perl version. See #748380 for details.

For this to work, packages containing binary perl modules need to migrate
from using the hardcoded /usr/lib/perl5 directory to the value of the
$Config{vendorarch} variable, as defined in the 'Config' module.

The implementation uses an executable debhelper install file.